ark_engine/requirements.txt
typer==0.9.0
rich==13.7.0

pypdf==4.0.1
ebooklib==0.18
beautifulsoup4==4.12.3

python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.2.0

Pillow==10.2.0
pytesseract==0.3.10

chardet==5.2.0
